Removal of ci_build
The reason for this is to have more control
over the text that comes with unstable builds.

The author of this project likes to provide different messages
based on the target audience.
The commit messages on GitHub are written for developers.
The Discord text for unstable builds are written for end users.

Creating a merge commit on unstable provides the chance
to combine multiple commits into one unstable build.