FMV: Sequence header configuration reset from outside
- Fixes regression with titles like Top Gun that
  have MPEG files with different settings

Note: For some reason, this state has wrong frame periods.
Tested using coneheads.
Note: For some reason, this state has also wrong frame periods when fast forwarding in Top Gun and resuming the playback.

Note: This is the pain. The linker scripts were wrong.
_end hat the same address as .noinit.seq_hdr_conf causing
the first variable - the frame period - to be reset along with other .bss symbols.
I keep this commit as a memento and not squash it.