VMPEG: Fixed image size registers 00E04002 and 00E04004
Written values were stored but could not be read back

Concerning "Addams Family" - Disc 2, it crashed when entering the menu.
This was caused by a "Division by Zero" exception, resetting the machine.
This was caused by corrupted data, caused by double execution of the instructions
at 0x0275ed0 (program of the movie player) which causes a zero value at 0xD01D8A.
This problem already manifests even before the movie starts playback, but the crash itself is only occuring when the value is read.
This happens only when the movie player menu is opened.

All of this was caused by the values in mentioned registers being not the written ones but artificial ones when the VMPEG verilog implementation was first conceived.

So, this change does:

- Fixes "Division by Zero" when opening movie player menu with "Addams Family" - Disc 2